Moved ST with not existing endpoint to DefaultTest
