+++ src/test/java/io/strimzi/kafka/bridge/http/services/ConsumerService.java
@@ -83,7 +83,6 @@ public class ConsumerService extends BaseService {
     }
 
     // Consumer actions
-
     public ConsumerService unsubscribeConsumer(VertxTestContext context, String groupId, String name, String... topicNames) throws InterruptedException, ExecutionException, TimeoutException {
         JsonArray topics = new JsonArray();
         for (String topicName : topicNames) {
